requirements.txt
gevent
flask
openai
python-dotenv
PyPDF2
langchain
faiss-cpu
openai-whisper
torch
Flask-Session
Flask-CORS